backend/requirements.txt
alembic==1.16.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pip==25.1.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.0
setuptools==80.9.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.0
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0